app/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
datetime==4.3
# debugging only
# ptvsd==4.3.2
